+++ src/client/BuildTreeVisualizer.svelte
