+++ src/client/FilerVisualizer.svelte
