+++ src/client/ServerVisualizer.svelte
