+++ src/client/bar.ts
