+++ src/client/foo.ts
@@ -1,0 +1,1 @@
+console.log('src/client/foo.ts says hi with import.meta.url:', import.meta.url);
