--- src/frontend/foo.ts
@@ -1,1 +1,0 @@
-console.log('src/frontend/foo.ts says hi with import.meta.url:', import.meta.url);
